+++ pl_bolts/models/self_supervised/amdim/__init__.py
@@ -1,6 +1,6 @@
 from pl_bolts.models.self_supervised.amdim.amdim_module import AMDIM
 from pl_bolts.models.self_supervised.amdim.networks import AMDIMEncoder
-from pl_bolts.models.self_supervised.amdim.transforms import (
+from pl_bolts.transforms.self_supervised.amdim_transforms import (
     AMDIMEvalTransformsCIFAR10,
     AMDIMEvalTransformsImageNet128,
     AMDIMEvalTransformsSTL10,
