+++ pl_bolts/models/self_supervised/amdim/datasets.py
@@ -3,7 +3,7 @@ from typing import Optional
 from torch.utils.data import random_split
 
 from pl_bolts.datasets import CIFAR10Mixed, UnlabeledImagenet
-from pl_bolts.models.self_supervised.amdim import transforms as amdim_transforms
+from pl_bolts.transforms.self_supervised import amdim_transforms as amdim_transforms
 from pl_bolts.utils import _TORCHVISION_AVAILABLE
 from pl_bolts.utils.stability import under_review
 from pl_bolts.utils.warnings import warn_missing_pkg
