+++ pl_bolts/models/self_supervised/simclr/__init__.py
@@ -1,4 +1,4 @@
-from pl_bolts.models.self_supervised.simclr.transforms import (  # noqa: F401
+from pl_bolts.transforms.self_supervised.simclr_transforms import (  # noqa: F401
     SimCLREvalDataTransform,
     SimCLRTrainDataTransform,
 )
