+++ pl_bolts/models/self_supervised/swav/__init__.py
@@ -1,7 +1,7 @@
 from pl_bolts.models.self_supervised.swav.loss import SWAVLoss
 from pl_bolts.models.self_supervised.swav.swav_module import SwAV
 from pl_bolts.models.self_supervised.swav.swav_resnet import resnet18, resnet50
-from pl_bolts.models.self_supervised.swav.transforms import (
+from pl_bolts.transforms.self_supervised.swav_transforms import (
     SwAVEvalDataTransform,
     SwAVFinetuneTransform,
     SwAVTrainDataTransform,
