+++ pl_bolts/models/self_supervised/swav/swav_finetuner.py
@@ -5,8 +5,8 @@ from pytorch_lightning import Trainer, seed_everything
 
 from pl_bolts.models.self_supervised.ssl_finetuner import SSLFineTuner
 from pl_bolts.models.self_supervised.swav.swav_module import SwAV
-from pl_bolts.models.self_supervised.swav.transforms import SwAVFinetuneTransform
 from pl_bolts.transforms.dataset_normalizations import imagenet_normalization, stl10_normalization
+from pl_bolts.transforms.self_supervised.swav_transforms import SwAVFinetuneTransform
 
 
 def cli_main():  # pragma: no cover
