+++ pl_bolts/transforms/self_supervised/moco_transforms.py
