+++ tests/models/self_supervised/unit/test_transforms.py
@@ -3,7 +3,7 @@ import pytest
 import torch
 from PIL import Image
 
-from pl_bolts.models.self_supervised.simclr.transforms import (
+from pl_bolts.transforms.self_supervised.simclr_transforms import (
     SimCLREvalDataTransform,
     SimCLRFinetuneTransform,
     SimCLRTrainDataTransform,
